conversion en uint64_t

+++ src/client/receiverwindow.cpp
@@ -16,7 +16,7 @@ void ReceiverInWindow::addProjectToList(ProjectData projet){
     window_->addProjectList(projet);
 }
 
-void ReceiverInWindow::updateProjectNameInList(long long id, const std::string& newName){
+void ReceiverInWindow::updateProjectNameInList(uint64_t id, const std::string& newName){
     window_->updateProjectNameInList(id, newName);
     
 }
